// src/contracts/types.ts
export type SupportedNetworkId = '421613'

export type ContractName =
  | 'BaseRegistrarImplementation'
  | 'ARBRegistrarController'
  | 'Multicall'
  | 'NameWrapper'
  | 'DNSRegistrar'
  | 'PublicResolver'
  | 'ENSRegistryWithFallback'
  | 'ReverseRegistrar'
  | 'UniversalResolver'
  | 'BulkRenewal'